+++ src/release/java/com/ryn/skyryn/dev/Dev.java
@@ -1,0 +1,7 @@
+package com.ryn.skyryn.dev;
+
+public final class Dev {
+	private Dev() { }
+
+	public static final boolean ON = false;
+}